move dumb/reusable components in the core folder, update README.md
