+++ README.md
@@ -1,22 +1,7 @@
-# Assignment-3
+# Assignment-4
 
-Hosted Project: https://capable-youtiao-eea16c.netlify.app/
 
-Assignment-3, Advancement over Assignment-2 Create a Dashboard/Landing Page to showcase "Top Gainers" & "Top Losers" in American Stock Market, also added Company's About page of all the listed company, dynamically routing through the pages, using API.
-
-Created a Dashboard component to showcase the result of the API.
-
-Dashboard utilised Indicator component which is reused for both "Top Gainers" & "Top Losers", with `data` and `type` being the only two props.
-
-Used Redux's provided `useSelector` & `dispatch` function to access data and call the API, using `fetchStock()` from actions `./src/actions/stockActions.js`.
-
-The `fetchStock()` function call the API, using axios. Simultenously, also managing to call the slice's reducers as `fetchStockRequest()`, `fetchStocksSuccess()` & `fetchStocksFailure()` defined in the `./src/reducers/stockReducer.js`.
-
-Slice is a combination of states and reducers, which helps tweaking the values of the state, and provide some action based on the state.
-
-After successfully creating slice, actions and reducer, now all of this can be connected to a redux-store. And now the redux store could be provided directly to the main parent component to supply one single source of data(ideology: single source of truth) as prop.
-
-Here, in `App.jsx` a `<Provider store={store}>` component is Wrapped around as a parent component to all the upcoming elements, supplying the store as prop to all upcomming element in the application.
+Assignment-4, Advancement over Assignment-3 Create a Product page linked to dashboard to showcase Various Products with their details[title, category, price, rating, stock] & also to fetch categories and filter the Table depending on selected filter/category, using API.
